Convert a duckplyr frame to a dbplyr table

This function converts a lazy duckplyr frame or a data frame to a dbplyr table in duckplyr's internal connection. This allows using dbplyr functions on the data, including hand-written SQL queries. Use as_duckdb_tibble() to convert back to a lazy duckplyr frame.

Usage

as_tbl(.data)

Arguments

.data

A lazy duckplyr frame or a data frame.

Value

A dbplyr table.

Examples

df <- duckdb_tibble(a = 1L)
df
#> # A duckplyr data frame: 1 variable
#>       a
#>   <int>
#> 1     1

tbl <- as_tbl(df)
tbl
#> # Source:   table<as_tbl_duckplyr_SwlKLUUEdL> [?? x 1]
#> # Database: DuckDB v1.2.2 [unknown@Linux 6.11.0-1013-azure:R 4.5.0//tmp/RtmpfUSrN4/duckplyr/duckplyr3bed33b41574.duckdb]
#>       a
#>   <int>
#> 1     1

tbl %>%
  mutate(b = sql("a + 1")) %>%
  as_duckdb_tibble()
#> # A duckplyr data frame: 2 variables
#>       a     b
#>   <int> <int>
#> 1     1     2