Compute results

This is a method for the dplyr::compute() generic. For a duckplyr frame, compute() executes a query but stores it in a (temporary) table, or in a Parquet or CSV file. The result is a duckplyr frame that can be used with subsequent dplyr verbs.

Usage

# S3 method for class 'duckplyr_df'
compute(
  x,
  ...,
  prudence = NULL,
  name = NULL,
  schema_name = NULL,
  temporary = TRUE
)

Arguments

x

A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...

Arguments passed on to methods

prudence

Either a string:

"frugal": a frugal data frame,

"lavish": a lavish data frame,

"thrifty": allow the materialization up to a maximum size of 1 million cells.

Or a named vector with at least one of

cells (numeric)

rows (numeric)

to allow materialization for data up to a certain size, measured in cells (values) and rows in the resulting data frame. The equivalent of "thrifty" is c(cells = 1e6).

If cells is specified but not rows, rows is Inf. If rows is specified but not cells, cells is Inf.

The default is to inherit the prudence of the input.

name

The name of the table to store the result in.

schema_name

The schema to store the result in, defaults to the current schema.

temporary

Set to FALSE to store the result in a permanent table.

See also

Examples

library(duckplyr)
df <- duckdb_tibble(x = c(1, 2))
df <- mutate(df, y = 2)
explain(df)
#> ┌───────────────────────────┐
#> │         PROJECTION        │
#> │    ────────────────────   │
#> │             x             │
#> │             y             │
#> │                           │
#> │          ~2 Rows          │
#> └─────────────┬─────────────┘
#> ┌─────────────┴─────────────┐
#> │     R_DATAFRAME_SCAN      │
#> │    ────────────────────   │
#> │         data.frame        │
#> │                           │
#> │       Projections: x      │
#> │                           │
#> │          ~2 Rows          │
#> └───────────────────────────┘
df <- compute(df)
explain(df)
#> ┌───────────────────────────┐
#> │         SEQ_SCAN          │
#> │    ────────────────────   │
#> │    duckplyr_SwlKLUUEdL    │
#> │                           │
#> │        Projections:       │
#> │             x             │
#> │             y             │
#> │                           │
#> │          ~2 Rows          │
#> └───────────────────────────┘