Compute results to a file

These functions apply to (lazy) duckplyr frames. They executes a query and stores the results in a flat file. The result is a duckplyr frame that can be used with subsequent dplyr verbs.

compute_parquet() creates a Parquet file.

compute_csv() creates a CSV file.

Usage

compute_parquet(x, path, ..., lazy = NULL, options = NULL)

compute_csv(x, path, ..., lazy = NULL, options = NULL)

Arguments

x

A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

path

The path to store the result in.

...

These dots are for future extensions and must be empty.

lazy

Set to TRUE to return a lazy or FALSE to return an eager data frame, see the "Eager and lazy" section in duckdb_tibble(). The default is to inherit the lazyness of the input.

options

A list of additional options to pass to create the storage format, see https://duckdb.org/docs/data/parquet/overview#writing-to-parquet-files or https://duckdb.org/docs/data/csv/overview#writing-using-the-copy-statement for details.

See also

Examples

library(duckplyr)
df <- data.frame(x = c(1, 2))
df <- mutate(df, y = 2)
path <- tempfile(fileext = ".parquet")
df <- compute_parquet(df, path)
explain(df)
#> ┌───────────────────────────┐
#> │       READ_PARQUET        │
#> │    ────────────────────   │
#> │         Function:         │
#> │        READ_PARQUET       │
#> │                           │
#> │        Projections:       │
#> │             x             │
#> │             y             │
#> │                           │
#> │          ~2 Rows          │
#> └───────────────────────────┘