duckplyr data frames

Data frames backed by duckplyr have a special class, "duckplyr_df", in addition to the default classes. This ensures that dplyr methods are dispatched correctly. For such objects, dplyr verbs such as dplyr::mutate(), dplyr::select() or dplyr::filter() will attempt to use DuckDB. If this is not possible, the original dplyr implementation is used.

duck_tbl() works like tibble::tibble(). In contrast to dbplyr, duckplyr data frames are "eager" by default. To avoid unwanted expensive computation, they can be converted to "lazy" duckplyr frames on which dplyr::collect() needs to be called explicitly.

as_duck_tbl() converts a data frame or a dplyr lazy table to a duckplyr data frame. This is a generic function that can be overridden for custom classes.

is_duck_tbl() returns TRUE if x is a duckplyr data frame.

Usage

duck_tbl(..., .lazy = FALSE)

as_duck_tbl(x, ..., .lazy = FALSE)

is_duck_tbl(x)

Arguments

...

For duck_tbl(), passed on to tibble::tibble(). For as_duck_tbl(), passed on to methods.

.lazy

Logical, whether to create a lazy duckplyr frame. If TRUE, dplyr::collect() must be called before the data can be accessed.

x

The object to convert or to test.

Value

For duck_tbl() and as_duck_tbl(), an object with the following classes:

"lazy_duckplyr_df" if .lazy is TRUE

"duckplyr_df"

Classes of a tibble::tibble

For is_duck_tbl(), a scalar logical.

Details

Set the DUCKPLYR_FALLBACK_INFO and DUCKPLYR_FORCE environment variables for more control over the behavior, see config for more details.

Examples

x <- duck_tbl(a = 1)
x
#> # A duckplyr data frame: 1 variable
#>       a
#>   <dbl>
#> 1     1

library(dplyr)
x %>%
  mutate(b = 2)
#> # A duckplyr data frame: 2 variables
#>       a     b
#>   <dbl> <dbl>
#> 1     1     2

x$a
#> [1] 1

y <- duck_tbl(a = 1, .lazy = TRUE)
y
#> # A duckplyr data frame: 1 variable
#>       a
#>   <dbl>
#> 1     1
try(length(y$a))
#> Error : Materialization is disabled, use collect() or as_tibble() to materialize
length(collect(y)$a)
#> [1] 1