duckplyr data frames

Data frames backed by duckplyr have a special class, "duckplyr_df", in addition to the default classes. This ensures that dplyr methods are dispatched correctly. For such objects, dplyr verbs such as dplyr::mutate(), dplyr::select() or dplyr::filter() will use DuckDB.

duckdb_tibble() works like tibble::tibble(), returning a lavish duckplyr data frame by default. See vignette("prudence") for details.

as_duckdb_tibble() converts a data frame or a dplyr lazy table to a duckplyr data frame. This is a generic function that can be overridden for custom classes.

is_duckdb_tibble() returns TRUE if x is a duckplyr data frame.

Usage

duckdb_tibble(..., .prudence = c("lavish", "thrifty", "frugal"))

as_duckdb_tibble(x, ..., prudence = c("lavish", "thrifty", "frugal"))

is_duckdb_tibble(x)

Arguments

...

For duckdb_tibble(), passed on to tibble::tibble(). For as_duckdb_tibble(), passed on to methods.

.prudence, prudence

Either a string:

"frugal": a frugal data frame,

"lavish": a lavish data frame,

"thrifty": allow the materialization up to a maximum size of 1 million cells.

Or a named vector with at least one of

cells (numeric)

rows (numeric)

to allow materialization for data up to a certain size, measured in cells (values) and rows in the resulting data frame. The equivalent of "thrifty" is c(cells = 1e6).

If cells is specified but not rows, rows is Inf. If rows is specified but not cells, cells is Inf.

The default is to inherit the prudence of the input.

x

The object to convert or to test.

Value

For duckdb_tibble() and as_duckdb_tibble(), an object with the following classes:

"prudent_duckplyr_df" if .prudence is not "lavish"

"duckplyr_df"

Classes of a tibble::tibble

For is_duckdb_tibble(), a scalar logical.

Examples

x <- duckdb_tibble(a = 1)
x
#> # A duckplyr data frame: 1 variable
#>       a
#>   <dbl>
#> 1     1

library(dplyr)
x %>%
  mutate(b = 2)
#> # A duckplyr data frame: 2 variables
#>       a     b
#>   <dbl> <dbl>
#> 1     1     2

x$a
#> [1] 1

y <- duckdb_tibble(a = 1, .prudence = "frugal")
y
#> # A duckplyr data frame: 1 variable
#>       a
#>   <dbl>
#> 1     1
try(length(y$a))
#> Error : Materialization is disabled, use collect() or as_tibble() to materialize.
length(collect(y)$a)
#> [1] 1