Class predicate for duckplyr data frames

Tests if the input object is of class "duckplyr_df".

Usage

is_duckplyr_df(.data)

Arguments

.data

The object to test

Value

TRUE if the input object is of class "duckplyr_df", otherwise FALSE.

Examples

tibble(a = 1:3) %>%
  is_duckplyr_df()
#> Warning: `is_duckplyr_df()` was deprecated in duckplyr 1.0.0.
#> ℹ Please use `is_duck_tbl()` instead.
#> [1] FALSE

tibble(a = 1:3) %>%
  as_duckplyr_df() %>%
  is_duckplyr_df()
#> [1] TRUE