Verbs not implemented in duckplyr

The following dplyr generics have no counterpart method in duckplyr. If you want to help add a new verb, please refer to our contributing guide https://duckplyr.tidyverse.org/CONTRIBUTING.html#support-new-verbs

Unsupported verbs

For these verbs, duckplyr will fall back to dplyr.

dplyr::add_count()

dplyr::compute()

dplyr::cross_join()

dplyr::do()

dplyr::group_by()

dplyr::group_indices()

dplyr::group_keys()

dplyr::group_map()

dplyr::group_modify()

dplyr::group_nest()

dplyr::group_size()

dplyr::group_split()

dplyr::group_trim()

dplyr::groups()

dplyr::n_groups()

dplyr::nest_by()

dplyr::nest_join()

dplyr::reframe()

dplyr::rename_with()

dplyr::rows_append()

dplyr::rows_delete()

dplyr::rows_insert()

dplyr::rows_patch()

dplyr::rows_update()

dplyr::rows_upsert()

dplyr::rowwise()

generics::setequal()

dplyr::slice_head()

dplyr::slice_sample()

dplyr::slice_tail()

dplyr::slice()

dplyr::ungroup()