Configuration options

The behavior of duckplyr can be fine-tuned with several environment variables, and one option.

Options

duckdb.materialize_message: Set to FALSE to turn off diagnostic output from duckdb on data frame materialization. Currenty set to TRUE when duckplyr is loaded.

Environment variables

DUCKPLYR_OUTPUT_ORDER: If TRUE, row output order is preserved. The default may change the row order where dplyr would keep it stable.

DUCKPLYR_FORCE: If TRUE, fail if duckdb cannot handle a request.

DUCKPLYR_FALLBACK_INFO: If TRUE, print a message when a fallback to dplyr occurs because duckdb cannot handle a request.

DUCKPLYR_CHECK_ROUNDTRIP: If TRUE, check if all columns are roundtripped perfectly when creating a relational object from a data frame, This is slow, and mostly useful for debugging. The default is to check roundtrip of attributes.

DUCKPLYR_EXPERIMENTAL: If TRUE, pass experimental = TRUE to certain duckdb functions. Currently unused.

Examples

# options(duckdb.materialize_message = FALSE)
data.frame(a = 3:1) %>%
  as_duckplyr_df() %>%
  inner_join(data.frame(a = 1:4), by = "a")
#> materializing:
#> ---------------------
#> --- Relation Tree ---
#> ---------------------
#> Projection [___coalesce(lhs.a_x, rhs.a_y) as a]
#>   Join REGULAR INNER ___eq_na_matches_na(lhs.a_x, rhs.a_y)
#>     Projection [a as a_x]
#>       r_dataframe_scan(0x55a19ae255f8)
#>     Projection [a as a_y]
#>       r_dataframe_scan(0x55a19ac9aab8)
#> 
#> ---------------------
#> -- Result Columns  --
#> ---------------------
#> - a (INTEGER)
#> 
#>   a
#> 1 1
#> 2 2
#> 3 3

rlang::with_options(duckdb.materialize_message = FALSE, {
  data.frame(a = 3:1) %>%
    as_duckplyr_df() %>%
    inner_join(data.frame(a = 1:4), by = "a") %>%
    print()
})
#>   a
#> 1 1
#> 2 2
#> 3 3

# Sys.setenv(DUCKPLYR_OUTPUT_ORDER = TRUE)
data.frame(a = 3:1) %>%
  as_duckplyr_df() %>%
  inner_join(data.frame(a = 1:4), by = "a")
#> materializing:
#> ---------------------
#> --- Relation Tree ---
#> ---------------------
#> Projection [___coalesce(lhs.a_x, rhs.a_y) as a]
#>   Join REGULAR INNER ___eq_na_matches_na(lhs.a_x, rhs.a_y)
#>     Projection [a as a_x]
#>       r_dataframe_scan(0x55a19ea49248)
#>     Projection [a as a_y]
#>       r_dataframe_scan(0x55a19ea4b868)
#> 
#> ---------------------
#> -- Result Columns  --
#> ---------------------
#> - a (INTEGER)
#> 
#>   a
#> 1 1
#> 2 2
#> 3 3

withr::with_envvar(c(DUCKPLYR_OUTPUT_ORDER = "TRUE"), {
  data.frame(a = 3:1) %>%
    as_duckplyr_df() %>%
    inner_join(data.frame(a = 1:4), by = "a")
})
#> materializing:
#> ---------------------
#> --- Relation Tree ---
#> ---------------------
#> Projection [___coalesce(lhs.a_x, rhs.a_y) as a]
#>   Order [lhs.___row_number_x ASC, rhs.___row_number_y ASC]
#>     Join REGULAR INNER ___eq_na_matches_na(lhs.a_x, rhs.a_y)
#>       Projection [a_x as a_x, row_number() OVER () as ___row_number_x]
#>         Projection [a as a_x]
#>           r_dataframe_scan(0x55a19ec8e6f8)
#>       Projection [a_y as a_y, row_number() OVER () as ___row_number_y]
#>         Projection [a as a_y]
#>           r_dataframe_scan(0x55a19eca8ab0)
#> 
#> ---------------------
#> -- Result Columns  --
#> ---------------------
#> - a (INTEGER)
#> 
#>   a
#> 1 3
#> 2 2
#> 3 1

# Sys.setenv(DUCKPLYR_FORCE = TRUE)
add_one <- function(x) {
  x + 1
}

data.frame(a = 3:1) %>%
  as_duckplyr_df() %>%
  mutate(b = add_one(a))
#>   a b
#> 1 3 4
#> 2 2 3
#> 3 1 2

try(withr::with_envvar(c(DUCKPLYR_FORCE = "TRUE"), {
  data.frame(a = 3:1) %>%
    as_duckplyr_df() %>%
    mutate(b = add_one(a))
}))
#> Error in do_translate(expr) : Unknown function: add_one

# Sys.setenv(DUCKPLYR_FALLBACK_INFO = TRUE)
withr::with_envvar(c(DUCKPLYR_FALLBACK_INFO = "TRUE"), {
  data.frame(a = 3:1) %>%
    as_duckplyr_df() %>%
    mutate(b = add_one(a))
})
#> Error processing with relational.
#> Caused by error in `do_translate()` at duckplyr/R/relational.R:274:3:
#> ! Unknown function: add_one
#> Requested fallback for relational:
#> ℹ No relational implementation for group_data()
#> Requested fallback for relational:
#> ℹ No relational implementation for dplyr_col_modify()
#>   a b
#> 1 3 4
#> 2 2 3
#> 3 1 2