duckplyr data frames

Data frames backed by duckplyr have a special class, "duckplyr_df", in addition to the default classes. This ensures that dplyr methods are dispatched correctly. For such objects, dplyr verbs such as dplyr::mutate(), dplyr::select() or dplyr::filter() will use DuckDB.

duckdb_tibble() works like tibble::tibble(), returning an "eager" duckplyr data frame by default. See the "Eager and lazy" section below.

as_duckdb_tibble() converts a data frame or a dplyr lazy table to a duckplyr data frame. This is a generic function that can be overridden for custom classes.

is_duckdb_tibble() returns TRUE if x is a duckplyr data frame.

Usage

duckdb_tibble(..., .lazy = FALSE)

as_duckdb_tibble(x, ..., .lazy = FALSE)

is_duckdb_tibble(x)

Arguments

...

For duckdb_tibble(), passed on to tibble::tibble(). For as_duckdb_tibble(), passed on to methods.

.lazy

Logical, whether to create a lazy duckplyr frame. See the section "Eager and lazy" for details.

x

The object to convert or to test.

Value

For duckdb_tibble() and as_duckdb_tibble(), an object with the following classes:

"lazy_duckplyr_df" if .lazy is TRUE

"duckplyr_df"

Classes of a tibble::tibble

For is_duckdb_tibble(), a scalar logical.

Eager and lazy

Data frames backed by duckplyr behave as regular data frames in almost all respects. In particular, direct column access like df$x, or retrieving the number of rows with nrow(), works identically. For a duckplyr frame that is the result of a dplyr operation, accessing column data or retrieving the number of rows will trigger a computation that is carried out by DuckDB.

Eager and lazy duckplyr frames differ in their behavior for column access and row count. For eager duckplyr frames, the underlying DuckDB computation is carried out upon the first request. Once the results are computed, they are cached and subsequent requests are fast. This is a good choice for small to medium-sized data, where DuckDB can provide a nice speedup but materializing the data is affordable. This is the default for duckdb_tibble() and as_duckdb_tibble(). Use .lazy = TRUE for these functions to check that all operations are supported by DuckDB.

For lazy duckplyr frames, accessing a column or requesting the number of rows triggers an error. This is a good choice for large data sets where the cost of materializing the data may be prohibitive due to size or computation time, and the user wants to control when the computation is carried out. This is the default for the ingestion functions like read_parquet_duckdb(). It is safe to use read_parquet_duckdb(lazy = FALSE) if the data is small enough to be materialized at any stage.

A lazy duckplyr frame can be converted to an eager one with as_duckdb_tibble(.lazy = FALSE). The collect.duckplyr_df() method converts to a plain tibble. Other useful methods include compute_file() for storing results in a file, and compute.duckplyr_df() for storing results in temporary storage on disk.

Computing via DuckDB is not always possible, see vignette("limits") for the supported operations. In such cases, the original dplyr implementation is used, see fallback for details. As the original dplyr implementation accesses columns directly, the data must be materialized before a fallback can be executed. This means that automatic fallback is only possible for "eager" duckplyr frames, while for "lazy" duckplyr frames, one of the aforementioned methods must be used.

The concept of lazy tables is also known from dbplyr, but "eager" tables are not available there at the time of writing, and the data must always be brought into R memory through dplyr::collect().

Examples

x <- duckdb_tibble(a = 1)
x
#> # A duckplyr data frame: 1 variable
#>       a
#>   <dbl>
#> 1     1

library(dplyr)
x %>%
  mutate(b = 2)
#> # A duckplyr data frame: 2 variables
#>       a     b
#>   <dbl> <dbl>
#> 1     1     2

x$a
#> [1] 1

y <- duckdb_tibble(a = 1, .lazy = TRUE)
y
#> # A duckplyr data frame: 1 variable
#>       a
#>   <dbl>
#> 1     1
try(length(y$a))
#> Error : Materialization is disabled, use collect() or as_tibble() to materialize
length(collect(y)$a)
#> [1] 1