Show stats

Prints statistics on how many calls were handled by DuckDB. The output shows the total number of requests in the current session, split by fallbacks to dplyr and requests handled by duckdb.

Usage

stats_show()

Value

Called for its side effect.

Examples

stats_show()
#> 🛠: 13
#> 🔨:  6
#> 🦆:  7
#> dplyr_col_modify, group_data, inner_join, mutate, mutate.data.frame

tibble(a = 1:3) %>%
  as_duckplyr_df() %>%
  mutate(b = a + 1)
#> materializing:
#> ---------------------
#> --- Relation Tree ---
#> ---------------------
#> Projection [a as a, +(a, 1.0) as b]
#>   r_dataframe_scan(0x55f3e126bd00)
#> 
#> ---------------------
#> -- Result Columns  --
#> ---------------------
#> - a (INTEGER)
#> - b (DOUBLE)
#> 
#> # A tibble: 3 × 2
#>       a     b
#>   <int> <dbl>
#> 1     1     2
#> 2     2     3
#> 3     3     4

stats_show()
#> 🛠: 14
#> 🔨:  6
#> 🦆:  8
#> dplyr_col_modify, group_data, inner_join, mutate, mutate.data.frame